requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
dj-database-url==2.3.0
Django==5.1.7
django-bootstrap5==23.4
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
mysqlclient==2.2.7
pillow==11.1.0
python-dotenv==1.1.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.2
